+++ world/src/area.rs
@@ -58,11 +58,11 @@ impl Encodable for Area {
 
 impl Area {
     pub fn new(rng_seed: u32, spec: AreaSpec) -> Area {
-        let num_mobs = 32u;
-        let num_items = 24u;
+        let num_mobs = 32;
+        let num_items = 24;
 
         let mut terrain = HashMap::new();
-        let mut rng: StdRng = SeedableRng::from_seed([rng_seed as uint + spec.depth as uint].as_slice());
+        let mut rng: StdRng = SeedableRng::from_seed([rng_seed as usize + spec.depth as usize].as_slice());
         mapgen::gen_herringbone(
             &mut rng,
             &spec,
@@ -88,14 +88,14 @@ impl Area {
         // XXX: copy-pasting the space-finding code.
         spawns.extend(
             action::random_spawns(
-                &mut rng, num_mobs, spec.depth as uint, spec.biome, Category::Mob)
+                &mut rng, num_mobs, spec.depth, spec.biome, Category::Mob)
             .into_iter().filter_map(|spawn|
             if let Some(loc) = opens.pop() { Some((spawn, loc))
             } else { None }));
 
         spawns.extend(
             action::random_spawns(
-                &mut rng, num_items, spec.depth as uint, spec.biome, Category::Item)
+                &mut rng, num_items, spec.depth as i32, spec.biome, Category::Item)
             .into_iter().filter_map(|spawn|
             if let Some(loc) = opens.pop() { Some((spawn, loc))
             } else { None }));
